requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
certifi==2018.11.29
cffi==1.14.1
chardet==3.0.4
configparser==5.0.0
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
google-images-download==2.8.0
html2markdown==0.1.7
html2text==2018.1.9
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
lxml==4.5.2
Markdown==3.2.2
markdown2==2.3.9
markdownify==0.5.1
mysql-connector-python==8.0.21
Naked==0.1.31
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-frontmatter==0.5.0
python-slugify==4.0.1
PyYAML==5.1
queuelib==1.5.0
requests==2.21.0
Scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
slugify==0.0.1
soupsieve==2.0.1
text-unidecode==1.3
tomd==0.1.3
toml==0.10.1
Twisted==20.3.0
url-normalize==1.4.2
urllib3==1.25
w3lib==1.22.0
zope.interface==5.1.0
